# src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: Cultural Navigator Routes API
  version: 1.0.0
  description: API для CRUD операций над маршрутами.
servers:
  - url: http://localhost:8080
paths:
  /api/routes:
    get:
      tags: [Routes]
      operationId: getAllRoutes
      summary: Получить список маршрутов
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RouteResponse'
    post:
      tags: [Routes]
      operationId: createRoute
      summary: Создать маршрут
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteRequest'
      responses:
        '201':
          description: Маршрут создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteResponse'
  /api/routes/{id}:
    get:
      tags: [Routes]
      operationId: getRouteById
      summary: Получить маршрут по id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteResponse'
        '404':
          description: Маршрут не найден
    put:
      tags: [Routes]
      operationId: updateRoute
      summary: Обновить маршрут
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteRequest'
      responses:
        '200':
          description: Маршрут обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteResponse'
        '404':
          description: Маршрут не найден
    delete:
      tags: [Routes]
      operationId: deleteRoute
      summary: Удалить маршрут
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Маршрут удален
        '404':
          description: Маршрут не найден
components:
  schemas:
    RouteRequest:
      type: object
      required: [title, authorId]
      properties:
        title:
          type: string
          minLength: 1
        authorId:
          type: integer
          format: int64
    RouteResponse:
      type: object
      required: [id, title, authorId, authorUsername, itemsCount]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        authorId:
          type: integer
          format: int64
        authorUsername:
          type: string
        itemsCount:
          type: integer
          format: int32